电路板图像部分除旋转归一化已完成
完成了综合形态匹配表的维护，以及报告的连接，与老师的图像处理部分的差错也完成了对接

+++ bzxt117/apps/evi/views.py
@@ -644,7 +644,10 @@ class devShapeEviViewset(viewsets.ModelViewSet):
         name = str(evi.srcImgURL).split("/")[-1]
         picType = os.path.splitext(name)[1]
         path = os.path.join(MEDIA_ROOT,"image/devShapeEvi/original/")
-        os.rename(os.path.join(MEDIA_ROOT,str(evi.srcImgURL)), os.path.join(path, str(id) + picType))
+        if os.path.exists(os.path.join(MEDIA_ROOT,str(evi.srcImgURL))):
+            os.rename(os.path.join(MEDIA_ROOT,str(evi.srcImgURL)), os.path.join(path, str(id) + picType))
+        else:
+            raise APIException("没上传电路板图片，请上传。")
         evi.srcImgURL = "image/devShapeEvi/original/" + str(id) + picType
         evi.save()
